requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.32.4
yfinance==0.2.65
openai==1.93.2
